InitTestModule refractor
