+++ packages/api/src/module/read/learning-materials/learning-materials.read-module.spec.ts
@@ -1,36 +1,18 @@
-import { Test, TestingModule } from '@nestjs/testing';
 import { LearningMaterials } from '@prisma/client';
 import { AsyncReturnType } from 'type-fest';
 import { v4 as uuid } from 'uuid';
 import waitForExpect from 'wait-for-expect';
 
-import { cleanupDatabase } from '@/common/test-utils';
+import { initTestModule } from '@/common/test-utils';
 import { LearningMaterialsUrlWasGenerated } from '@/events/learning-materials-url-was-generated.domain-event';
 import { ApplicationEvent } from '@/module/application-command-events';
-import { PrismaService } from '@/prisma/prisma.service';
 import { UserId } from '@/users/users.types';
-import { ApplicationEventBus } from '@/write/shared/application/application.event-bus';
 import { EventStreamName } from '@/write/shared/application/event-stream-name.value-object';
-
-import { AppModule } from '../../../app.module';
 
 const SAMPLE_MATERIALS_URL = 'https://app.process.st/runs/jNMTGn96H8Xe3H8DbcpJOg';
 
 async function learningMaterialsTestModule() {
-  const app: TestingModule = await Test.createTestingModule({
-    imports: [AppModule],
-  }).compile();
-
-  await app.init();
-
-  const eventBus = app.get<ApplicationEventBus>(ApplicationEventBus);
-  const prismaService = app.get<PrismaService>(PrismaService);
-
-  await cleanupDatabase(prismaService);
-
-  function eventOccurred(event: ApplicationEvent): void {
-    eventBus.publishAll([event]);
-  }
+  const { prismaService, close, eventOccurred } = await initTestModule();
 
   async function expectReadModel(expectation: { courseUserId: UserId; readModel: LearningMaterials | null }) {
     await waitForExpect(() =>
@@ -38,10 +20,6 @@ async function learningMaterialsTestModule() {
         prismaService.learningMaterials.findUnique({ where: { courseUserId: expectation.courseUserId } }),
       ).resolves.toStrictEqual(expectation.readModel),
     );
-  }
-
-  async function close() {
-    await app.close();
   }
 
   return { eventOccurred, expectReadModel, close };
